【modify】 直播首推观看人数接口   首推ID 修改为整形

+++ zyx-mapper/src/main/java/com/zyx/mapper/account/AccountInfoMapper.java
@@ -22,5 +22,5 @@ public interface AccountInfoMapper extends BaseMapper<AccountInfo> {
 
     int renewpwd(UserLoginParam userLoginParam);
 
-    List<AccountInfoVo> queryUserDevaluation(List<Long> users);
+    List<AccountInfoVo> queryUserDevaluation(List<Integer> users);
 }
